Treat nil variables as empty dictionary.
Before v0.2 URIWithVariables: (now deprecated) accepted a nil argument and treated it as an empty dictionary of variables. This behavior was undocumented and untested, but was relied upon by a dependent project.

This commit restores the nil variables behavior to the deprecated URIWithVariables: method. It does not change the behavior of the replacement relativeStringWithVariables:error: method.